+++ oaaclient/tests/test_client.py
@@ -6,8 +6,10 @@ license that can be found in the LICENSE file or at
 https://opensource.org/licenses/MIT.
 """
 
-from requests import delete
-from unittest.mock import patch
+from requests.exceptions import JSONDecodeError as RequestsJSONDecodeError
+from requests.models import Response
+from unittest.mock import MagicMock, patch
+import json
 import os
 import pytest
 import uuid
@@ -120,7 +122,6 @@ def test_client_data_source(veza_con):
     veza_con.delete_provider(provider_id)
 
 
-# @patch("oaaclient.client.get_provider_list")
 @pytest.mark.parametrize("url",["https://noreply.vezacloud.com", "noreply.vezacloud.com", "noreply.vezacloud.com/", "https://noreply.vezacloud.com/"])
 def test_url_formatter(url):
     test_api_key = "1234"
@@ -128,3 +129,226 @@ def test_url_formatter(url):
         veza_con = OAAClient(url=url, token=test_api_key)
 
         assert veza_con.url == "https://noreply.vezacloud.com"
+
+@patch('oaaclient.client.requests')
+def test_api_get_error(mock_requests):
+    # Test that the correct OAAClient exception is raised on properly populated
+
+    test_api_key = "1234"
+    # patch get_provider_list to instantiate a connection object
+    with patch.object(OAAClient, "get_provider_list", return_value=[]):
+        veza_con = OAAClient(url="https://noreply.vezacloud.com", token=test_api_key)
+
+
+    mock_response = Response()
+    mock_response.status_code = 400
+    error_message = b"""
+                    {
+                        "code": "Internal",
+                        "message": "Internal Server Error, please retry and if the error persists contact support at support@veza.com",
+                        "request_id": "2271c08a9abd3b425c88a397b01bb351",
+                        "timestamp": "2022-08-05T21:12:29.405153171Z",
+                        "details": [
+                            {
+                                "@type": "type.googleapis.com/errorstatus.v1.UserFacingErrorInfo",
+                                "reason": "INTERNAL",
+                                "metadata": {},
+                                "message": "Internal server error.",
+                                "resolution": "Please retry and if the error persists contact support at support@veza.com"
+                            }
+                        ]
+                    }
+                    """
+    mock_response._content = error_message
+    mock_requests.get.return_value = mock_response
+
+    with pytest.raises(OAAClientError) as e:
+        veza_con.api_get("/api/path")
+
+    # test that the error is populated propery
+    assert e.value.error == "Internal"
+    assert e.value.message == "Internal Server Error, please retry and if the error persists contact support at support@veza.com"
+    assert e.value.status_code == 400
+    assert len(e.value.details) == 1
+    assert "Internal server error." in str(e.value.details)
+
+@patch('oaaclient.client.requests')
+def test_api_get_nonjson_error(mock_requests):
+    # Test that the OAAClient correctly handles a non-JSON respponse if error isn't coming from Veza stack
+
+    test_api_key = "1234"
+    url = "https://noreply.vezacloud.com"
+    # patch get_provider_list to instantiate a connection object
+    with patch.object(OAAClient, "get_provider_list", return_value=[]):
+        veza_con = OAAClient(url=url, token=test_api_key)
+
+    # Mock a response with non-JSON data, will force a JSONDecodeError
+    mock_response = Response()
+    mock_response.status_code = 500
+    mock_response._content = b"This is not json"
+    mock_response.reason = "Error Reason"
+    mock_response.url = url
+
+    mock_requests.get.return_value = mock_response
+
+    with pytest.raises(OAAClientError) as e:
+        veza_con.api_get("/api/path")
+
+    # should recieve the generic error message
+    assert e.value.error == "ERROR"
+    assert "Error Reason" in e.value.message
+    assert e.value.status_code == 500
+
+
+@patch('oaaclient.client.requests')
+def test_api_post_error(mock_requests):
+    # Test that the correct OAAClient exception is raised on properly populated
+
+    test_api_key = "1234"
+    # patch get_provider_list to instantiate a connection object
+    with patch.object(OAAClient, "get_provider_list", return_value=[]):
+        veza_con = OAAClient(url="https://noreply.vezacloud.com", token=test_api_key)
+
+
+    mock_response = Response()
+    mock_response.status_code = 400
+    error_message = b"""
+            {
+                "code": "InvalidArgument",
+                "message": "Invalid Arguments",
+                "request_id": "1091d23a67ad44a63723fc050280e5ae",
+                "timestamp": "2022-08-05T19:59:11.508388808Z",
+                "details": [
+                    {
+                    "@type": "type.googleapis.com/google.rpc.BadRequest",
+                    "field_violations": [
+                        {
+                        "field": "name",
+                        "description": "Provider with the same name already exists"
+                        }
+                    ]
+                    },
+                    {
+                    "@type": "type.googleapis.com/errorstatus.v1.UserFacingErrorInfo",
+                    "reason": "INVALID_ARGUMENTS",
+                    "metadata": {},
+                    "message": "Request includes invalid arguments.",
+                    "resolution": "Reference error details for the exact field violations."
+                    }
+                ]
+            }
+            """
+    mock_response._content = error_message
+
+    mock_requests.post.return_value = mock_response
+
+    with pytest.raises(OAAClientError) as e:
+        veza_con.api_post("/api/path", data={})
+
+    # test that the error is populated propery
+    assert e.value.error == "InvalidArgument"
+    assert e.value.message == "Invalid Arguments"
+    assert e.value.status_code == 400
+    assert e.value.details != []
+    assert "Provider with the same name already exists" in str(e.value.details)
+
+@patch('oaaclient.client.requests')
+def test_api_post_nonjson_error(mock_requests):
+    # Test that the OAAClient correctly handles a non-JSON respponse if error isn't coming from Veza stack
+
+    test_api_key = "1234"
+    url = "https://noreply.vezacloud.com"
+    # patch get_provider_list to instantiate a connection object
+    with patch.object(OAAClient, "get_provider_list", return_value=[]):
+        veza_con = OAAClient(url=url, token=test_api_key)
+
+    # Mock a response with non-JSON data, will force a JSONDecodeError
+    mock_response = Response()
+    mock_response.status_code = 500
+    mock_response._content = b"This is not json"
+    mock_response.reason = "Error Reason"
+    mock_response.url = url
+
+    mock_requests.post.return_value = mock_response
+
+    with pytest.raises(OAAClientError) as e:
+        veza_con.api_post("/api/path", data={})
+
+    # should recieve the generic error message
+    assert e.value.error == "ERROR"
+    assert "Error Reason" in e.value.message
+    assert e.value.status_code == 500
+
+@patch('oaaclient.client.requests')
+def test_api_delete_error(mock_requests):
+    # Test that the correct OAAClient exception is raised on properly populated
+
+    test_api_key = "1234"
+    # patch get_provider_list to instantiate a connection object
+    with patch.object(OAAClient, "get_provider_list", return_value=[]):
+        veza_con = OAAClient(url="https://noreply.vezacloud.com", token=test_api_key)
+
+
+    mock_response = Response()
+    mock_response.status_code = 404
+    error_message = b"""
+            {
+                "code": "NotFound",
+                "message": "Not Found",
+                "request_id": "1de5e43499c90f2036cdfe92ed76f58e",
+                "timestamp": "2022-08-05T21:06:53.046972349Z",
+                "details": [
+                    {
+                    "@type": "type.googleapis.com/errorstatus.v1.ResourceInfo",
+                    "resource_type": "datasource",
+                    "resource": "b1e654e7-2104-4180-9dee-2f76e2b52463"
+                    },
+                    {
+                    "@type": "type.googleapis.com/errorstatus.v1.UserFacingErrorInfo",
+                    "reason": "NOT_FOUND",
+                    "metadata": {},
+                    "message": "Requested resource was not found.",
+                    "resolution": ""
+                    }
+                ]
+            }
+            """
+    mock_response._content = error_message
+    mock_requests.delete.return_value = mock_response
+
+    with pytest.raises(OAAClientError) as e:
+        veza_con.api_delete("/api/path")
+
+    # test that the error is populated propery
+    assert e.value.error == "NotFound"
+    assert e.value.message == "Not Found"
+    assert e.value.status_code == 404
+    assert e.value.details != []
+    assert "Requested resource was not found." in str(e.value.details)
+
+@patch('oaaclient.client.requests')
+def test_api_post_delete_error(mock_requests):
+    # Test that the OAAClient correctly handles a non-JSON respponse if error isn't coming from Veza stack
+
+    test_api_key = "1234"
+    url = "https://noreply.vezacloud.com"
+    # patch get_provider_list to instantiate a connection object
+    with patch.object(OAAClient, "get_provider_list", return_value=[]):
+        veza_con = OAAClient(url=url, token=test_api_key)
+
+    # Mock a response with non-JSON data, will force a JSONDecodeError
+    mock_response = Response()
+    mock_response.status_code = 500
+    mock_response._content = b"This is not json"
+    mock_response.reason = "Error Reason"
+    mock_response.url = url
+
+    mock_requests.delete.return_value = mock_response
+
+    with pytest.raises(OAAClientError) as e:
+        veza_con.api_delete("/api/path")
+
+    # should recieve the generic error message
+    assert e.value.error == "ERROR"
+    assert "Error Reason" in e.value.message
+    assert e.value.status_code == 500